combined hook + map
