+++ src/components/Map/Map.tsx
@@ -1,6 +1,7 @@
 import React, { useRef, useState, useEffect } from "react";
 import { Wrapper, Status } from "@googlemaps/react-wrapper";
-import { useDeepCompareEffectForMaps } from "./hooks/DeepCompare/DeepCompare"
+import { createCustomEqual } from "fast-equals";
+import { isLatLngLiteral } from "@googlemaps/typescript-guards";
 
 interface MapProps extends google.maps.MapOptions {
   style: { [key: string]: string };
@@ -61,5 +62,43 @@ const Map: React.FC<MapProps> = ({
     </>
   );
 };
+
+const deepCompareEqualsForMaps = createCustomEqual(
+  (deepEqual) => (a: any, b: any) => {
+    if (
+      isLatLngLiteral(a) ||
+      a instanceof google.maps.LatLng ||
+      isLatLngLiteral(b) ||
+      b instanceof google.maps.LatLng
+    ) {
+      return new google.maps.LatLng(a).equals(new google.maps.LatLng(b));
+    }
+
+    // to-do extend to other types
+    // use fast-equals for other objects
+    return deepEqual(a, b);
+  }
+);
+
+function useDeepCompareMemoize(value: any) {
+  const ref = React.useRef();
+
+  if (!deepCompareEqualsForMaps(value, ref.current)) {
+    ref.current = value;
+  }
+
+  return ref.current;
+}
+
+function useDeepCompareEffectForMaps(
+  callback: React.EffectCallback,
+  dependencies: any[]
+) {
+  React.useEffect(callback, dependencies.map(useDeepCompareMemoize));
+}
+
+// window.addEventListener("DOMContentLoaded", () => {
+//   ReactDom.render(<App />, document.getElementById("root"));
+// });
 
 export default Map;
