--- src/components/Map/hooks/DeepCompare/DeepCompare.tsx
@@ -1,48 +1,0 @@
-import React, { useEffect } from "react";
-import * as ReactDom from "react-dom";
-import { createCustomEqual } from "fast-equals";
-import { isLatLngLiteral } from "@googlemaps/typescript-guards";
-
-const deepCompareEqualsForMaps = createCustomEqual(
-  (deepEqual) => (a: any, b: any) => {
-    if (
-      isLatLngLiteral(a) ||
-      a instanceof google.maps.LatLng ||
-      isLatLngLiteral(b) ||
-      b instanceof google.maps.LatLng
-    ) {
-      return new google.maps.LatLng(a).equals(new google.maps.LatLng(b));
-    }
-
-    // to-do extend to other types
-    // use fast-equals for other objects
-    return deepEqual(a, b);
-  }
-);
-
-function useDeepCompareMemoize(value: any) {
-  const ref = React.useRef();
-
-  if (!deepCompareEqualsForMaps(value, ref.current)) {
-    ref.current = value;
-  }
-
-  return ref.current;
-}
-
-function useDeepCompareEffectForMaps(
-  callback: React.EffectCallback,
-  dependencies: any[]
-) {
-  React.useEffect(callback, dependencies.map(useDeepCompareMemoize));
-}
-
-// window.addEventListener("DOMContentLoaded", () => {
-//   ReactDom.render(<App />, document.getElementById("root"));
-// });
-
-export {
-  deepCompareEqualsForMaps,
-  useDeepCompareEffectForMaps,
-  useDeepCompareMemoize
-}
